+++ src/plugins/world_control/WorldControl_TEST.cc
@@ -35,7 +35,7 @@ using namespace gui;
 
 // See https://github.com/ignitionrobotics/ign-gui/issues/75
 /////////////////////////////////////////////////
-TEST(WorldControlTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(Load))
+TEST(WorldControlTest, IGN_UTILS_TEST_ENABLED_ONLY_ON_LINUX(Load))
 {
   common::Console::SetVerbosity(4);
 
@@ -60,7 +60,7 @@ TEST(WorldControlTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(Load))
 }
 
 /////////////////////////////////////////////////
-TEST(WorldControlTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(WorldControl))
+TEST(WorldControlTest, IGN_UTILS_TEST_ENABLED_ONLY_ON_LINUX(WorldControl))
 {
   common::Console::SetVerbosity(4);
 
@@ -122,111 +122,6 @@ TEST(WorldControlTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(WorldControl))
   // Play
   plugin->OnPlay();
   EXPECT_TRUE(playCalled);
-
-  // Cleanup
-  plugins.clear();
-}
-
-/////////////////////////////////////////////////
-TEST(WorldControlTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(WorldNameNoService))
-{
-  common::Console::SetVerbosity(4);
-
-  Application app(g_argc, g_argv);
-  app.AddPluginPath(std::string(PROJECT_BINARY_PATH) + "/lib");
-
-  // Get main window
-  auto win = app.findChild<MainWindow *>();
-  ASSERT_NE(nullptr, win);
-
-  QStringList names{"banana", "grape"};
-  win->setProperty("worldNames", names);
-
-  // Load plugin
-  const char *pluginStr =
-    "<plugin filename=\"WorldControl\">"
-    "</plugin>";
-
-  tinyxml2::XMLDocument pluginDoc;
-  EXPECT_EQ(tinyxml2::XML_SUCCESS, pluginDoc.Parse(pluginStr));
-  EXPECT_TRUE(app.LoadPlugin("WorldControl",
-      pluginDoc.FirstChildElement("plugin")));
-
-  // Show, but don't exec, so we don't block
-  win->QuickWindow()->show();
-
-  // Get plugin
-  auto plugins = win->findChildren<plugins::WorldControl *>();
-  EXPECT_EQ(plugins.size(), 1);
-
-  // World control service
-  bool pauseCalled = false;
-  std::function<bool(const msgs::WorldControl &, msgs::Boolean &)> cb =
-      [&](const msgs::WorldControl &_req, msgs::Boolean &)
-  {
-    pauseCalled = _req.pause();
-    return true;
-  };
-  transport::Node node;
-  node.Advertise("/world/banana/control", cb);
-
-  // Pause
-  plugins[0]->OnPause();
-  EXPECT_TRUE(pauseCalled);
-
-  // Cleanup
-  plugins.clear();
-}
-
-/////////////////////////////////////////////////
-TEST(WorldControlTest, IGN_UTILS_TEST_DISABLED_ON_WIN32(WorldNameBadService))
-{
-  common::Console::SetVerbosity(4);
-
-  Application app(g_argc, g_argv);
-  app.AddPluginPath(std::string(PROJECT_BINARY_PATH) + "/lib");
-
-  // Get main window
-  auto win = app.findChild<MainWindow *>();
-  ASSERT_NE(nullptr, win);
-
-  QStringList names{"banana", "grape"};
-  win->setProperty("worldNames", names);
-
-  // Load plugin
-  const char *pluginStr =
-    "<plugin filename=\"WorldControl\">"
-    "  <service>/world/watermelon/control</service>"
-    "</plugin>";
-
-  tinyxml2::XMLDocument pluginDoc;
-  EXPECT_EQ(tinyxml2::XML_SUCCESS, pluginDoc.Parse(pluginStr));
-  EXPECT_TRUE(app.LoadPlugin("WorldControl",
-      pluginDoc.FirstChildElement("plugin")));
-
-  // Show, but don't exec, so we don't block
-  win->QuickWindow()->show();
-
-  // Get plugin
-  auto plugins = win->findChildren<plugins::WorldControl *>();
-  EXPECT_EQ(plugins.size(), 1);
-
-  // World control service
-  bool pauseCalled = false;
-  std::function<bool(const msgs::WorldControl &, msgs::Boolean &)> cb =
-      [&](const msgs::WorldControl &_req, msgs::Boolean &)
-  {
-    pauseCalled = _req.pause();
-    return true;
-  };
-  transport::Node node;
-
-  // banana, not watermelon
-  node.Advertise("/world/banana/control", cb);
-
-  // Pause
-  plugins[0]->OnPause();
-  EXPECT_TRUE(pauseCalled);
 
   // Cleanup
   plugins.clear();